fix print warning at compile time on ubuntu.

+++ ndarray/pygpu_ndarray_object.h
@@ -73,17 +73,17 @@ void PyGpuNdArray_fprint(FILE * fd, const PyGpuNdArrayObject *self)
 {
     fprintf(fd, "PyGpuNdArrayObject <%p, %p> nd=%i data_allocated=%d\n",
 	    self, PyGpuNdArray_DATA(self), PyGpuNdArray_NDIM(self), self->data_allocated);
-    fprintf(fd, "\tRefcount: %d\n", self->ob_refcnt);
+    fprintf(fd, "\tRefcount: %ld\n", (long int)self->ob_refcnt);
     fprintf(fd, "\tBASE: %p\n", PyGpuNdArray_BASE(self));
     fprintf(fd, "\tHOST_DIMS:      ");
     for (int i = 0; i < PyGpuNdArray_NDIM(self); ++i)
     {
-        fprintf(fd, "%i\t", PyGpuNdArray_DIMS(self)[i]);
+        fprintf(fd, "%ld\t", PyGpuNdArray_DIMS(self)[i]);
     }
     fprintf(fd, "\n\tHOST_STRIDES: ");
     for (int i = 0; i < PyGpuNdArray_NDIM(self); ++i)
     {
-        fprintf(fd, "%i\t", PyGpuNdArray_STRIDES(self)[i]);
+        fprintf(fd, "%ld\t", PyGpuNdArray_STRIDES(self)[i]);
     }
     fprintf(fd, "\n\tFLAGS: ");
     fprintf(fd, "\n\t\tC_CONTIGUOUS: %d", PyGpuNdArray_ISCONTIGUOUS(self));
@@ -105,12 +105,12 @@ void PyArray_fprint(FILE * fd, const PyArrayObject *self)
     fprintf(fd, "\tHOST_DIMS:      ");
     for (int i = 0; i < PyArray_NDIM(self); ++i)
     {
-        fprintf(fd, "%i\t", PyArray_DIMS(self)[i]);
+        fprintf(fd, "%ld\t", PyArray_DIMS(self)[i]);
     }
     fprintf(fd, "\n\tHOST_STRIDES: ");
     for (int i = 0; i < PyArray_NDIM(self); ++i)
     {
-        fprintf(fd, "%i\t", PyArray_STRIDES(self)[i]);
+        fprintf(fd, "%ld\t", PyArray_STRIDES(self)[i]);
     }
     fprintf(fd, "\n\tFLAGS: ");
     fprintf(fd, "\n\t\tC_CONTIGUOUS: %d", PyArray_ISCONTIGUOUS(self));
